Fix all -Werror warnings for GHC 9.6.7 CI
- Remove unused imports across all modules (Data.List, Data.Set, Data.Map, etc.)
- Fix type-default warnings with explicit type annotations
- Fix name-shadowing warning in Extract.hs tag parsing
- Replace partial  with safe  in tests
- Suppress orphan instance warnings in test spec
- Remove overlapping pattern in Main.hs path display
- Fix redundant imports in FGL, CommunityGraph, HTML, Obsidian, etc.

+++ src/Graphos/UseCase/Benchmark.hs
@@ -5,8 +5,6 @@ module Graphos.UseCase.Benchmark
   , printBenchmark
   ) where
 
-import qualified Data.Text as T
-
 -- | Benchmark result
 data BenchmarkResult = BenchmarkResult
   { brCorpusWords    :: Int
@@ -34,4 +32,4 @@ printBenchmark result = do
   putStrLn "\n=== Token Reduction Benchmark ==="
   putStrLn $ "Full corpus:     " ++ show (brCorpusWords result) ++ " words (~" ++ show (brCorpusWords result `div` 250) ++ " tokens)"
   putStrLn $ "Typical subgraph: ~" ++ show (brSubgraphTokens result) ++ " tokens"
-  putStrLn $ "Token savings:    " ++ show (round (brSavingsPercent result :: Double)) ++ "%"
+  putStrLn $ "Token savings:    " ++ show (round (brSavingsPercent result) :: Int) ++ "%"
